moved some methods to event.dart, database takes event obj instead
