+++ lib/temp_frontend2.dart
@@ -2,6 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:planner/common/database.dart';
+import 'package:planner/models/event.dart';
 
 var auth = FirebaseAuth.instanceFor(
     app: Firebase.app(), persistence: Persistence.LOCAL);
@@ -73,11 +74,11 @@ class MyAppState extends State<MyApp> {
     if (event == "add") {
       final timeStart = DateTime.now();
       final timeEnd = timeStart.add(const Duration(hours: 8));
-      await db.addUniqueUserEvent(
-          eventName: "example_event_name_2",
-          eventTags: {"example_event_tag1"},
+      await db.addUniqueUserEvent(Event(
+          name: "example_event_name_2",
+          tags: {"example_event_tag1"},
           timeStart: timeStart,
-          timeEnd: timeEnd);
+          timeEnd: timeEnd));
       return;
     } else if (event == "get range") {
       final dateStart = DateTime.parse("2023-10-20");
@@ -161,14 +162,14 @@ class MyAppState extends State<MyApp> {
                           hintText: 'Add a event here...',
                           enabledBorder: OutlineInputBorder(
                             borderRadius:
-                            BorderRadius.all(Radius.circular(12.0)),
+                                BorderRadius.all(Radius.circular(12.0)),
                             borderSide: BorderSide(color: Colors.red, width: 2),
                           ),
                           focusedBorder: OutlineInputBorder(
                             borderRadius:
-                            BorderRadius.all(Radius.circular(12.0)),
+                                BorderRadius.all(Radius.circular(12.0)),
                             borderSide:
-                            BorderSide(color: Colors.red, width: 1.5),
+                                BorderSide(color: Colors.red, width: 1.5),
                           ),
                         ),
                       ),
